+++ mirrord/config/src/lib.rs
@@ -299,6 +299,16 @@ pub struct LayerConfig {
     /// # internal_proxy {#root-internal_proxy}
     #[config(nested)]
     pub internal_proxy: InternalProxyConfig,
+
+    /// ## use_proxy {#root-use_proxy}
+    ///
+    /// When disabled, mirrord will remove `HTTP[S]_PROXY` env variables before
+    /// doing any network requests. This is useful when the system sets a proxy
+    /// but you don't want mirrord to use it.
+    /// This also applies to the mirrord process (as it just removes the env).
+    /// If the remote pod sets this env, the mirrord process will still use it.
+    #[config(env = "MIRRORD_PROXY", default = true)]
+    pub use_proxy: bool,
 }
 
 impl LayerConfig {
@@ -491,7 +501,7 @@ impl LayerFileConfig {
         match path.as_ref().extension().and_then(|os_val| os_val.to_str()) {
             Some("json") => Ok(serde_json::from_str::<Self>(&rendered)?),
             Some("toml") => Ok(toml::from_str::<Self>(&rendered)?),
-            Some("yaml") => Ok(serde_yaml::from_str::<Self>(&rendered)?),
+            Some("yaml" | "yml") => Ok(serde_yaml::from_str::<Self>(&rendered)?),
             _ => Err(ConfigError::UnsupportedFormat),
         }
     }
@@ -744,6 +754,7 @@ mod tests {
             sip_binaries: None,
             kube_context: None,
             internal_proxy: None,
+            use_proxy: None,
         };
 
         assert_eq!(config, expect);